+++ core/event-service/src/main/java/ru/practicum/service/CompilationServiceImpl.java
@@ -71,8 +71,8 @@ public class CompilationServiceImpl implements CompilationService {
 		Compilation compilation = CompilationMapper.toEntity(compilationDto, events);
 		Compilation savedCompilation = compilationRepository.save(compilation);
 
-		Map<Long, Long> confirmedRequests = new HashMap<>();
-		savedCompilation.getEvents().forEach(event -> confirmedRequests.put(event.getId(), 0L));
+		Map<Long, Integer> confirmedRequests = new HashMap<>();
+		savedCompilation.getEvents().forEach(event -> confirmedRequests.put(event.getId(), 0));
 
 		Map<Long, Long> views = new HashMap<>();
 		savedCompilation.getEvents().forEach(event -> views.put(event.getId(), 0L));
@@ -131,7 +131,7 @@ public class CompilationServiceImpl implements CompilationService {
 			return Collections.emptyList();
 		}
 
-		Map<Long, Long> allConfirmedRequests = getConfirmedRequests(compilations);
+		Map<Long, Integer> allConfirmedRequests = getConfirmedRequests(compilations);
 		Map<Long, Long> allViews = getViews(compilations);
 
 		// Теперь маппим, просто подставляя готовые Map
@@ -153,7 +153,7 @@ public class CompilationServiceImpl implements CompilationService {
 
 	/// Map<eventId, confirmedRequests>
 	@NonNull
-	private Map<Long, Long> getConfirmedRequests(Collection<Compilation> compilations) {
+	private Map<Long, Integer> getConfirmedRequests(Collection<Compilation> compilations) {
 		List<Event> events = compilations.stream()
 				.flatMap(c -> c.getEvents().stream())
 				.toList();
